compilation errors

+++ hw2/hw2/PercolationStats.java
@@ -11,7 +11,7 @@ public class PercolationStats {
         if (N <= 0 || T <= 0) {
             throw new IllegalArgumentException();
         }
-        xs = new int[T];
+        xs = new double[T];
         this.T = T;
         for (int i = 0; i < T; i++) {
             Percolation p = pf.make(N);
